+++ server/service/endpoint_labels.go
@@ -7,84 +7,6 @@ import (
 	"github.com/go-kit/kit/endpoint"
 )
 
-type getLabelRequest struct {
-	ID uint
-}
-
-type labelResponse struct {
-	fleet.Label
-	DisplayText string `json:"display_text"`
-	Count       int    `json:"count"`
-	HostIDs     []uint `json:"host_ids"`
-}
-
-type getLabelResponse struct {
-	Label labelResponse `json:"label"`
-	Err   error         `json:"error,omitempty"`
-}
-
-func (r getLabelResponse) error() error { return r.Err }
-
-func labelResponseForLabel(ctx context.Context, svc fleet.Service, label *fleet.Label) (*labelResponse, error) {
-	return &labelResponse{
-		Label:       *label,
-		DisplayText: label.Name,
-		Count:       label.HostCount,
-	}, nil
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Get Label
-////////////////////////////////////////////////////////////////////////////////
-
-func makeGetLabelEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(getLabelRequest)
-		label, err := svc.GetLabel(ctx, req.ID)
-		if err != nil {
-			return getLabelResponse{Err: err}, nil
-		}
-		resp, err := labelResponseForLabel(ctx, svc, label)
-		if err != nil {
-			return getLabelResponse{Err: err}, nil
-		}
-		return getLabelResponse{Label: *resp}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Modify Label
-////////////////////////////////////////////////////////////////////////////////
-
-type modifyLabelRequest struct {
-	ID      uint
-	payload fleet.ModifyLabelPayload
-}
-
-type modifyLabelResponse struct {
-	Label labelResponse `json:"label"`
-	Err   error         `json:"error,omitempty"`
-}
-
-func (r modifyLabelResponse) error() error { return r.Err }
-
-func makeModifyLabelEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(modifyLabelRequest)
-		label, err := svc.ModifyLabel(ctx, req.ID, req.payload)
-		if err != nil {
-			return modifyLabelResponse{Err: err}, nil
-		}
-
-		labelResp, err := labelResponseForLabel(ctx, svc, label)
-		if err != nil {
-			return modifyLabelResponse{Err: err}, nil
-		}
-
-		return modifyLabelResponse{Label: *labelResp}, err
-	}
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 // List Labels
 ////////////////////////////////////////////////////////////////////////////////
