+++ server/service/handler.go
@@ -68,8 +68,6 @@ type FleetEndpoints struct {
 	SubmitLogs                            endpoint.Endpoint
 	CarveBegin                            endpoint.Endpoint
 	CarveBlock                            endpoint.Endpoint
-	ModifyLabel                           endpoint.Endpoint
-	GetLabel                              endpoint.Endpoint
 	ListLabels                            endpoint.Endpoint
 	ListHostsInLabel                      endpoint.Endpoint
 	DeleteLabel                           endpoint.Endpoint
@@ -154,8 +152,6 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 		GetQuerySpec:                          authenticatedUser(svc, makeGetQuerySpecEndpoint(svc)),
 		CreateDistributedQueryCampaign:        authenticatedUser(svc, makeCreateDistributedQueryCampaignEndpoint(svc)),
 		CreateDistributedQueryCampaignByNames: authenticatedUser(svc, makeCreateDistributedQueryCampaignByNamesEndpoint(svc)),
-		ModifyLabel:                           authenticatedUser(svc, makeModifyLabelEndpoint(svc)),
-		GetLabel:                              authenticatedUser(svc, makeGetLabelEndpoint(svc)),
 		ListLabels:                            authenticatedUser(svc, makeListLabelsEndpoint(svc)),
 		ListHostsInLabel:                      authenticatedUser(svc, makeListHostsInLabelEndpoint(svc)),
 		DeleteLabel:                           authenticatedUser(svc, makeDeleteLabelEndpoint(svc)),
@@ -242,8 +238,6 @@ type fleetHandlers struct {
 	SubmitLogs                            http.Handler
 	CarveBegin                            http.Handler
 	CarveBlock                            http.Handler
-	ModifyLabel                           http.Handler
-	GetLabel                              http.Handler
 	ListLabels                            http.Handler
 	ListHostsInLabel                      http.Handler
 	DeleteLabel                           http.Handler
@@ -322,8 +316,6 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		SubmitLogs:                            newServer(e.SubmitLogs, decodeSubmitLogsRequest),
 		CarveBegin:                            newServer(e.CarveBegin, decodeCarveBeginRequest),
 		CarveBlock:                            newServer(e.CarveBlock, decodeCarveBlockRequest),
-		ModifyLabel:                           newServer(e.ModifyLabel, decodeModifyLabelRequest),
-		GetLabel:                              newServer(e.GetLabel, decodeGetLabelRequest),
 		ListLabels:                            newServer(e.ListLabels, decodeListLabelsRequest),
 		ListHostsInLabel:                      newServer(e.ListHostsInLabel, decodeListHostsInLabelRequest),
 		DeleteLabel:                           newServer(e.DeleteLabel, decodeDeleteLabelRequest),
@@ -567,8 +559,6 @@ func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 	r.Handle("/api/v1/fleet/queries/run", h.CreateDistributedQueryCampaign).Methods("POST").Name("create_distributed_query_campaign")
 	r.Handle("/api/v1/fleet/queries/run_by_names", h.CreateDistributedQueryCampaignByNames).Methods("POST").Name("create_distributed_query_campaign_by_names")
 
-	r.Handle("/api/v1/fleet/labels/{id:[0-9]+}", h.ModifyLabel).Methods("PATCH").Name("modify_label")
-	r.Handle("/api/v1/fleet/labels/{id:[0-9]+}", h.GetLabel).Methods("GET").Name("get_label")
 	r.Handle("/api/v1/fleet/labels", h.ListLabels).Methods("GET").Name("list_labels")
 	r.Handle("/api/v1/fleet/labels/{id:[0-9]+}/hosts", h.ListHostsInLabel).Methods("GET").Name("list_hosts_in_label")
 	r.Handle("/api/v1/fleet/labels/{name}", h.DeleteLabel).Methods("DELETE").Name("delete_label")
@@ -673,6 +663,8 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, opts []kitht
 	e.POST("/api/v1/fleet/hosts/{id:[0-9]+}/refetch", refetchHostEndpoint, refetchHostRequest{})
 
 	e.POST("/api/v1/fleet/labels", createLabelEndpoint, createLabelRequest{})
+	e.PATCH("/api/v1/fleet/labels/{id:[0-9]+}", modifyLabelEndpoint, modifyLabelRequest{})
+	e.GET("/api/v1/fleet/labels/{id:[0-9]+}", getLabelEndpoint, getLabelRequest{})
 
 	e.GET("/api/v1/fleet/queries/run", runLiveQueryEndpoint, runLiveQueryRequest{})
