+++ server/service/labels.go
@@ -68,3 +68,103 @@ func (svc *Service) NewLabel(ctx context.Context, p fleet.LabelPayload) (*fleet.
 	}
 	return label, nil
 }
+
+////////////////////////////////////////////////////////////////////////////////
+// Modify Label
+////////////////////////////////////////////////////////////////////////////////
+
+type modifyLabelRequest struct {
+	ID uint `json:"-" url:"id"`
+	fleet.ModifyLabelPayload
+}
+
+type modifyLabelResponse struct {
+	Label labelResponse `json:"label"`
+	Err   error         `json:"error,omitempty"`
+}
+
+func (r modifyLabelResponse) error() error { return r.Err }
+
+func modifyLabelEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*modifyLabelRequest)
+	label, err := svc.ModifyLabel(ctx, req.ID, req.ModifyLabelPayload)
+	if err != nil {
+		return modifyLabelResponse{Err: err}, nil
+	}
+
+	labelResp, err := labelResponseForLabel(ctx, svc, label)
+	if err != nil {
+		return modifyLabelResponse{Err: err}, nil
+	}
+
+	return modifyLabelResponse{Label: *labelResp}, err
+}
+
+func (svc *Service) ModifyLabel(ctx context.Context, id uint, payload fleet.ModifyLabelPayload) (*fleet.Label, error) {
+	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionWrite); err != nil {
+		return nil, err
+	}
+
+	label, err := svc.ds.Label(ctx, id)
+	if err != nil {
+		return nil, err
+	}
+	if payload.Name != nil {
+		label.Name = *payload.Name
+	}
+	if payload.Description != nil {
+		label.Description = *payload.Description
+	}
+	return svc.ds.SaveLabel(ctx, label)
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// Get Label
+////////////////////////////////////////////////////////////////////////////////
+
+type getLabelRequest struct {
+	ID uint `url:"id"`
+}
+
+type labelResponse struct {
+	fleet.Label
+	DisplayText string `json:"display_text"`
+	Count       int    `json:"count"`
+	HostIDs     []uint `json:"host_ids"`
+}
+
+type getLabelResponse struct {
+	Label labelResponse `json:"label"`
+	Err   error         `json:"error,omitempty"`
+}
+
+func (r getLabelResponse) error() error { return r.Err }
+
+func getLabelEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*getLabelRequest)
+	label, err := svc.GetLabel(ctx, req.ID)
+	if err != nil {
+		return getLabelResponse{Err: err}, nil
+	}
+	resp, err := labelResponseForLabel(ctx, svc, label)
+	if err != nil {
+		return getLabelResponse{Err: err}, nil
+	}
+	return getLabelResponse{Label: *resp}, nil
+}
+
+func (svc *Service) GetLabel(ctx context.Context, id uint) (*fleet.Label, error) {
+	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionRead); err != nil {
+		return nil, err
+	}
+
+	return svc.ds.Label(ctx, id)
+}
+
+func labelResponseForLabel(ctx context.Context, svc fleet.Service, label *fleet.Label) (*labelResponse, error) {
+	return &labelResponse{
+		Label:       *label,
+		DisplayText: label.Name,
+		Count:       label.HostCount,
+	}, nil
+}
