+++ server/service/labels_test.go
@@ -1,0 +1,44 @@
+package service
+
+import (
+	"context"
+	"testing"
+
+	"github.com/fleetdm/fleet/v4/server/datastore/mysql"
+	"github.com/fleetdm/fleet/v4/server/fleet"
+	"github.com/fleetdm/fleet/v4/server/test"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestLabelsWithDS(t *testing.T) {
+	ds := mysql.CreateMySQLDS(t)
+
+	cases := []struct {
+		name string
+		fn   func(t *testing.T, ds *mysql.Datastore)
+	}{
+		{"GetLabel", testLabelsGetLabel},
+	}
+	for _, c := range cases {
+		t.Run(c.name, func(t *testing.T) {
+			defer mysql.TruncateTables(t, ds)
+			c.fn(t, ds)
+		})
+	}
+}
+
+func testLabelsGetLabel(t *testing.T, ds *mysql.Datastore) {
+	svc := newTestService(ds, nil, nil)
+
+	label := &fleet.Label{
+		Name:  "foo",
+		Query: "select * from foo;",
+	}
+	label, err := ds.NewLabel(context.Background(), label)
+	assert.Nil(t, err)
+	assert.NotZero(t, label.ID)
+
+	labelVerify, err := svc.GetLabel(test.UserContext(test.UserAdmin), label.ID)
+	assert.Nil(t, err)
+	assert.Equal(t, label.ID, labelVerify.ID)
+}
